tests/wpt/css-tests/css-shapes-1_dev/html/shape-outside-invalid-circle-000.htm

<!DOCTYPE html>
<html><head>
    <meta charset="utf-8">
    <title>CSS Shapes Tests: Circle - invalid 2 args</title>
    <link href="biqing.kwok@qq.com" rel="author" title="biqing">
    <link href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes" rel="help">
    <link href="http://www.w3.org/TR/css-shapes-1/#funcdef-circle" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link href="/resources/testharness.css" rel="stylesheet">
    <meta content="dom" name="flags">
    <meta content="shape-outside declaration is invalid if circle shape function only has two parameters" name="assert">
    <style type="text/css">
    #float {
        shape-outside: circle(50px, 50px);
        float: left;
        width: 100px;
        height: 100px;
    }
    </style>
</head>
<body>
    <div id="log"></div>

    <div id="float"></div>

    <script>

    test(function() {
        assert_true(getComputedStyle(document.getElementById('float')).shapeOutside == 'none');
    }, 'Circle - invalid 2 args');

    </script>

</body></html>
